VideoFERwebApp/requirements.txt
-i https://pypi.org/simple
absl-py==0.10.0
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
flask==1.1.2
gast==0.4.0
google-auth-oauthlib==0.4.1
google-auth==1.15.0
google-pasta==0.2.0
grpcio==1.37.0
gunicorn
h5py==3.1.0
idna==2.9
importlib-metadata==1.6.0 ; python_version < '3.8'
ipykernel==5.2.1
ipython-genutils==0.2.0
ipython==7.14.0
itsdangerous==1.1.0
jedi==0.17.0
jinja2==2.11.2
json5==0.9.4
jsonschema==3.2.0
#jupyter-client==6.1.3
#jupyter-core==4.6.3
#jupyterlab-server==1.1.4
#jupyterlab==2.1.2
keras-preprocessing==1.1.2
kiwisolver==1.2.0
livelossplot==0.5.0
markdown==3.2.2
markupsafe==1.1.1
#matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
#notebook==6.0.3
numpy==1.19.3
oauthlib==3.1.0
opencv-python-headless==4.3.0.38
opt-einsum==3.3.0
packaging==20.4
#pandas==1.1.5
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0 ; sys_platform != 'win32'
pickleshare==0.7.5
pillow==7.1.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.17.3
ptyprocess==0.6.0 ; os_name != 'nt'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
requests-oauthlib==1.3.0
requests==2.23.0
rsa==4.0
scipy==1.7.3
#seaborn==0.10.1
send2trash==1.5.0
six==1.15.0
tensorboard-plugin-wit==1.6.0.post3
tensorboard==2.6.0
tensorflow-estimator==2.6.0
tensorflow-cpu==2.6.2
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
werkzeug==1.0.1
wheel==0.35.0
wrapt==1.12.1
zipp==3.1.0